chore: add build script and move sample files
